Fix warnings and deprications
Generally this will make running the tests and using resque easier since
the console isn't flooded with warnings.

- Overshadowing variables in templates
  - Warning because you might use the wrong variable.
- Redefinition of constants
  - Warning because it might not be what you intended.
- Unused variables (e.g. _foo)
  - Warning because you might expect them to be used.
- Usage of undefined instance variables
  - Warning because doing so could be because of a typo.

There are still warnings but they are harder to fix. For example, the
warning about setting a global value on Encoding should probably not
be hidden and we don't know how to work around it right now.

The monkeypatch override of UTF handling in 1.9 can't really be fixed
without a redesign.

And probably a few other things as well.